backend/requirements.txt
flask>=3.0.0
flask-cors
python-dotenv
google-generativeai
PyMuPDF>=1.24.0
spacy>=3.0.0
